refactor(blocks): Markdown エディタの write/preview シェルを共通化する (#501)
BlockMarkdownInput と MarkdownFieldInput で逐語重複していた tabClass・タブバー・プレビュー枠・空表示を MarkdownEditorShell に集約する。compact/comfortable サイズで block インスペクタ用と全文フィールド用を切替え、write ペイン（textarea・任意ツールバー）は children として注入。見た目・挙動は不変。
