+++ frontend/src/features/edit-entity-text-fields/ui/MarkdownFieldInput.tsx
@@ -1,7 +1,5 @@
-import { useCallback, useRef, useState } from 'react'
-import { useTranslation } from '@/shared/i18n'
-import { Stack, Text } from '@/shared/ui'
-import { PublicMarkdownContent } from '@/shared/ui/markdown/PublicMarkdownContent'
+import { useCallback, useRef } from 'react'
+import { MarkdownEditorShell } from './MarkdownEditorShell'
 
 interface MarkdownFieldInputProps {
   id: string
@@ -121,8 +119,6 @@ export function MarkdownFieldInput({
   disabled,
   onChange,
 }: MarkdownFieldInputProps) {
-  const { t } = useTranslation()
-  const [tab, setTab] = useState<'write' | 'preview'>('write')
   const textareaRef = useRef<HTMLTextAreaElement>(null)
 
   const applyToolbar = useCallback(
@@ -141,91 +137,38 @@ export function MarkdownFieldInput({
     [value, onChange],
   )
 
-  const tabClass = (active: boolean) =>
-    [
-      'pb-1.5 font-sans text-caption font-medium transition-colors focus-visible:outline-none',
-      active
-        ? 'border-b-2 border-accent text-text-primary'
-        : 'text-text-muted hover:text-text-primary',
-    ].join(' ')
-
   return (
-    <Stack gap="xs">
-      <label htmlFor={id} className="font-sans text-body font-medium text-text-primary">
-        {label}
-      </label>
-
-      <div className="rounded-md border border-border shadow-sm">
-        {/* ── Tab bar ── */}
-        <div className="flex items-center justify-between border-b border-border px-inline-sm">
-          <div className="flex gap-inline-md">
-            <button
-              type="button"
-              onClick={() => {
-                setTab('write')
-              }}
-              className={tabClass(tab === 'write')}
-            >
-              {t('admin.markdownEditor.write')}
-            </button>
-            <button
-              type="button"
-              onClick={() => {
-                setTab('preview')
-              }}
-              className={tabClass(tab === 'preview')}
-            >
-              {t('admin.markdownEditor.preview')}
-            </button>
-          </div>
-        </div>
-
-        {/* ── Toolbar (write モードのみ) ── */}
-        {tab === 'write' && (
-          <div className="flex flex-wrap gap-0.5 border-b border-border bg-surface-overlay px-inline-sm py-stack-xs">
-            {TOOLBAR_ACTIONS.map((action) => (
-              <button
-                key={action.label}
-                type="button"
-                title={action.label}
-                disabled={disabled}
-                onClick={() => {
-                  applyToolbar(action)
-                }}
-                className="flex w-8 items-center justify-center rounded px-1.5 py-0.5 font-mono text-caption text-text-muted transition-colors hover:bg-surface-raised hover:text-text-primary disabled:cursor-not-allowed disabled:opacity-40"
-              >
-                {action.icon}
-              </button>
-            ))}
-          </div>
-        )}
-
-        {/* ── Write pane ── */}
-        {tab === 'write' && (
-          <textarea
-            ref={textareaRef}
-            id={id}
-            value={value}
+    <MarkdownEditorShell id={id} label={label} value={value} size="comfortable">
+      {/* ── Toolbar ── */}
+      <div className="flex flex-wrap gap-0.5 border-b border-border bg-surface-overlay px-inline-sm py-stack-xs">
+        {TOOLBAR_ACTIONS.map((action) => (
+          <button
+            key={action.label}
+            type="button"
+            title={action.label}
             disabled={disabled}
-            rows={16}
-            onChange={(e) => {
-              onChange(e.target.value)
+            onClick={() => {
+              applyToolbar(action)
             }}
-            className="w-full rounded-b-md bg-surface-raised px-inline-md py-stack-sm font-mono text-body text-text-primary focus-visible:outline-none disabled:cursor-not-allowed disabled:opacity-50"
-          />
-        )}
-
-        {/* ── Preview pane ── */}
-        {tab === 'preview' && (
-          <div className="min-h-64 rounded-b-md bg-surface-raised px-inline-md py-stack-sm">
-            {value.trim() === '' ? (
-              <Text muted>{t('admin.markdownEditor.empty')}</Text>
-            ) : (
-              <PublicMarkdownContent markdown={value} />
-            )}
-          </div>
-        )}
+            className="flex w-8 items-center justify-center rounded px-1.5 py-0.5 font-mono text-caption text-text-muted transition-colors hover:bg-surface-raised hover:text-text-primary disabled:cursor-not-allowed disabled:opacity-40"
+          >
+            {action.icon}
+          </button>
+        ))}
       </div>
-    </Stack>
+
+      {/* ── Write pane ── */}
+      <textarea
+        ref={textareaRef}
+        id={id}
+        value={value}
+        disabled={disabled}
+        rows={16}
+        onChange={(e) => {
+          onChange(e.target.value)
+        }}
+        className="w-full rounded-b-md bg-surface-raised px-inline-md py-stack-sm font-mono text-body text-text-primary focus-visible:outline-none disabled:cursor-not-allowed disabled:opacity-50"
+      />
+    </MarkdownEditorShell>
   )
 }
